refactor(general): :art: moved related components into their folders
