+++ src/features/coinView/views/CoinView.tsx
@@ -1,8 +1,8 @@
 import { Flex } from 'antd';
 
 import CoinsDropdown from '../components/CoinsDropdown';
-import AskBidWidget from '../components/AskBidWidget';
-import LadderWidget from '../components/LadderWidget';
+import AskBidWidget from '../components/bestAskAndBid/AskBidWidget';
+import LadderWidget from '../components/orderBook/LadderWidget';
 import ChartWidget from '../components/ChartWidget';
 
 function CoinView() {
